+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
-## v0.1 - 2023-4-16
+## v0.2 - 2023-04-20
+
+- Fixed issue with AI allies not telling you where they landing in some circumstances
+- Support an AI ally with multiple commanders landing on multiple worlds
+- Allies inform you when they're colonising a new planet
+
+## v0.1 - 2023-04-16
 
 - AI allies will notify you which planet they landed on on multi-planet maps